Extract interface from Episode number extractor
