+++ src/main/java/org/atlasapi/remotesite/btvod/BtVodModule.java
@@ -355,7 +355,7 @@ public class BtVodModule {
     }
     
     public DerivingFromSeriesBrandImageExtractor brandImageExtractor(String baseUrl) {
-        return new DerivingFromSeriesBrandImageExtractor(brandUriExtractor(baseUrl), seriesUriExtractor(baseUrl), itemImageExtractor(), new BtVodEpisodeNumberExtractor(mpxVodClient(baseUrl, btVodMpxProdFeedQParam)));
+        return new DerivingFromSeriesBrandImageExtractor(brandUriExtractor(baseUrl), seriesUriExtractor(baseUrl), itemImageExtractor(), new BtVodMpxBackedEpisodeNumberExtractor(mpxVodClient(baseUrl, btVodMpxProdFeedQParam)));
     }
     
     public ImageExtractor itemImageExtractor() {
